# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kolibri AI OS V2.1",
    "version": "2.1.0"
  },
  "paths": {
...
    "/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_v1_projects_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "operationId": "create_project_v1_projects_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Patch Project",
        "operationId": "patch_project_v1_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "operationId": "delete_project_v1_projects__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/restore": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Restore Project",
        "operationId": "restore_project_v1_projects__project_id__restore_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/messages": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Messages",
        "operationId": "list_messages_v1_projects__project_id__messages_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Add Message",
        "operationId": "add_message_v1_projects__project_id__messages_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
...
    "/v1/plans": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Plans",
        "operationId": "plans_v1_plans_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/subagents": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Subagents",
        "operationId": "subagents_v1_subagents_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/approvals": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Approvals",
        "operationId": "approvals_v1_approvals_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "Create Approval",
        "operationId": "create_approval_v1_approvals_post",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/canvases": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Canvases",
        "operationId": "canvases_v1_canvases_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/canvases/{canvas_id}": {
      "patch": {
        "tags": [
          "governance"
        ],
        "summary": "Patch Canvas",
        "operationId": "patch_canvas_v1_canvases__canvas_id__patch",
        "parameters": [
          {
            "name": "canvas_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Canvas Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanvasPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/formulalm/traces": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "Formula Trace",
        "operationId": "formula_trace_v1_formulalm_traces_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormulaTraceCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CanvasPatch": {
        "properties": {
          "state": {
            "additionalProperties": true,
            "type": "object",
            "title": "State"
          }
        },
        "type": "object",
        "title": "CanvasPatch"
      },
...
      "FormulaTraceCreate": {
        "properties": {
          "provenance": {
            "type": "string",
            "title": "Provenance"
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "default": false
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "provenance"
        ],
        "title": "FormulaTraceCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageCreate": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "default": "user"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MessageCreate"
      },
...
      "ProjectCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "default": "Новый проект"
          }
        },
        "type": "object",
        "title": "ProjectCreate"
      },
      "ProjectPatch": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ProjectPatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "sk-kolibri API key or signed Kolibri session",
        "description": "OpenAI-compatible bearer authentication."
      },
      "KolibriSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kolibri_session",
        "description": "Secure HttpOnly browser session cookie."
      },
      "KolibriSessionHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kolibri-Session"
      },
      "KolibriNodeToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kolibri-Node-Token"
      }
    }
  },
...
}